Champion
-Don't apply LOD for objects with armatures
-Set LOD strategy to 'screen_ratio_pixel_count'
-Write LOD meshes to 'LOD' sub-folder
-Updated 'lodDistance' property to Float so pixel area values can be provided

+++ ui/export.py
@@ -192,7 +192,7 @@ class _OgreCommonExport_(object):
         description="MESH number of LOD levels",
         min=0, max=32,
         default=config.get('lodLevels'))
-    EX_lodDistance = IntProperty(
+    EX_lodDistance = FloatProperty(
         name="LOD Distance",
         description="MESH distance increment to reduce LOD",
         min=0, max=2000, default=config.get('lodDistance'))
